+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 
 ### Features
+- Support password recovery using SMTP
 - Support Dask from Jupyter notebook
 
 ## 0.1.0